spring-bit-ml/requirements.txt
fastapi
aiokafka
uvicorn
pydantic
pydantic-settings
python-dotenv